pyproject.toml
[project]
name = "voxsupport"
version = "0.1.0"
description = "Production-grade open-source voice customer-support agent"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11,<3.13"
authors = [{ name = "Pratik Narendra Raut" }]

dependencies = [
    # Voice orchestration
    "pipecat-ai[daily,silero,openai,anthropic,deepgram]>=0.0.50",

    # STT / TTS — Kyutai stack (streaming, semantic VAD)
    # Install separately: pip install moshi (for Kyutai STT/TTS via delayed-streams-modeling)
    "moshi>=0.2.0",

    # LLM providers — LiteLLM for multi-provider routing + failover
    "litellm>=1.50.0",
    "openai>=1.50.0",
    "anthropic>=0.40.0",

    # Agent + state machine + MCP
    "langchain>=0.3.0",
    "langgraph>=0.2.40",
    "mcp>=0.9.0",

    # RAG
    "pgvector>=0.3.0",
    "psycopg[binary,pool]>=3.2.0",
    "sentence-transformers>=3.0.0",

    # Memory
    "redis>=5.1.0",
    "sqlalchemy[asyncio]>=2.0.0",

    # Guardrails
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "llm-guard>=0.3.0",

    # Observability
    "langfuse>=2.50.0",
    "prometheus-client>=0.20.0",

    # Web transport + server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=13.0",

    # Telephony (optional, week 5)
    "twilio>=9.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "ipykernel>=6.29.0",
]

benchmarks = [
    "pandas>=2.2.0",
    "matplotlib>=3.9.0",
    "soundfile>=0.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
